Added explicit whitespace handling to some rules.
All rules are now marked with their whitespace handling policy:
'!' marks rules that allow for implicit whitespace consumption.
'@' and '$' mark rules that don't allow implicit whitespace consumption.
'_' are for rules that are discarded by the parser, they allow implicit whitespace consumption,
but this may not always be desirable. Extra care must be taken with these rules.