Removed lifetime specifier from 'OutputTarget' trait.
Even though most _implementations_ of this trait will have lifetime specifiers (because they borrow things),
it's incorrect to _require_ a lifetime on the trait itself. For instance, it's possible for an implementation to
not borrow any underlying buffer (maybe it holds an owned one). What would the lifetime be here?

For 'OutputTarget' it's pretty useless anyways. Because you never look into the target's underlying memory.