Removed lifetime specifier from 'InputSource' trait.
Even though most _implementations_ of this trait will have lifetime specifiers (because they borrow things),
it's incorrect to _require_ a lifetime on the trait itself. For instance, it's possible for an implementation to
not borrow any underlying buffer (maybe it holds an owned one). What would the lifetime be here?

This does have a negative impact on the API's usability. Now, you cannot call a function from the API while holding
data returned from another call. This shouldn't be a common need in practice though.
And not all implementations would be able to uphold the old behavior anyways. For example, an implementation
that sat over a buffer couldn't guarantee the buffer would be unmutated between function calls.